+++ src/Validator/NotEmpty.php
@@ -17,25 +17,14 @@
  */
 namespace Mcustiel\SimpleRequest\Validator;
 
-use Mcustiel\SimpleRequest\Interfaces\ValidatorInterface;
-
 /**
  * Checks that the given value is not empty, based in the
  * <a href="http://php.net/manual/en/function.empty.php">empty definition of PHP.</a>
  *
  * @author mcustiel
  */
-class NotEmpty implements ValidatorInterface
+class NotEmpty  extends AbstractEmptySpecificationValidator
 {
-    /**
-     * {@inheritdoc}
-     *
-     * @see \Mcustiel\SimpleRequest\Interfaces\Specificable::setSpecification()
-     */
-    public function setSpecification($specification = null)
-    {
-    }
-
     /**
      * {@inheritdoc}
      *
