+++ src/Validator/UniqueItems.php
@@ -17,24 +17,13 @@
  */
 namespace Mcustiel\SimpleRequest\Validator;
 
-use Mcustiel\SimpleRequest\Interfaces\ValidatorInterface;
-
 /**
  * Checks if all the elements of a given array are unique values.
  *
  * @author mcustiel
  */
-class UniqueItems implements ValidatorInterface
+class UniqueItems extends AbstractEmptySpecificationValidator
 {
-    /**
-     * {@inheritdoc}
-     *
-     * @see \Mcustiel\SimpleRequest\Interfaces\Specificable::setSpecification()
-     */
-    public function setSpecification($specification = null)
-    {
-    }
-
     /**
      * {@inheritdoc}
      *
